third_party/WebKit/LayoutTests/virtual/serviceworker/http/tests/serviceworker/activation-after-registration-expected.txt


FAIL activation occurs after registration assert_equals: worker should be in the "parsed" state upon registration expected "parsed" but got "installing"
